+++ src/components/UI/Records/features/TransactionFormFields/index.ts
@@ -1,0 +1,1 @@
+export { TransactionFormFields } from './TransactionFormFields';